バッファサイズ関連を unsigned へ変更
INT32_MAX より大きい値を扱う時に比較が正しく行われないのを回避する為に変更。
サーバ側より極端に大きい(異常な)値を正しくはじくのが目的。
演算でのオーバーフローの考慮はちゃんと行う必要があるので、別途チェックを強化する。

git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@6711 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ ttssh2/ttxssh/buffer.h
@@ -7,9 +7,9 @@
 
 typedef struct buffer {
 	char *buf;      /* バッファの先頭ポインタ。realloc()により変動する。*/
-	int offset;     /* 現在の読み出し位置 */
-	int maxlen;     /* バッファの最大サイズ */
-	int len;        /* バッファに含まれる有効なデータサイズ */
+	size_t offset;     /* 現在の読み出し位置 */
+	size_t maxlen;     /* バッファの最大サイズ */
+	size_t len;        /* バッファに含まれる有効なデータサイズ */
 } buffer_t;
 
 /* buffer_t.buf の拡張の上限値 (16MB) */
@@ -20,7 +20,7 @@ typedef struct buffer {
 void buffer_clear(buffer_t *buf);
 buffer_t *buffer_init(void);
 void buffer_free(buffer_t *buf);
-void *buffer_append_space(buffer_t * buf, int size);
+void *buffer_append_space(buffer_t * buf, size_t size);
 int buffer_append(buffer_t *buf, char *ptr, int size);
 int buffer_append_length(buffer_t *msg, char *ptr, int size);
 void buffer_put_raw(buffer_t *msg, char *ptr, int size);
@@ -42,13 +42,13 @@ void buffer_get_ecpoint(char **data, const EC_GROUP *curve, EC_POINT *point);
 void buffer_get_ecpoint_msg(buffer_t *msg, const EC_GROUP *curve, EC_POINT *point);
 char *buffer_tail_ptr(buffer_t *msg);
 int buffer_overflow_verify(buffer_t *msg, int len);
-void buffer_consume(buffer_t *buf, int shift_byte);
-void buffer_consume_end(buffer_t *buf, int shift_byte);
+void buffer_consume(buffer_t *buf, size_t shift_byte);
+void buffer_consume_end(buffer_t *buf, size_t shift_byte);
 int buffer_compress(z_stream *zstream, char *payload, int len, buffer_t *compbuf);
 int buffer_decompress(z_stream *zstream, char *payload, int len, buffer_t *compbuf);
-int buffer_get_ret(buffer_t *msg, void *buf, int len);
-int buffer_get_int_ret(int *ret, buffer_t *msg);
-int buffer_get_int(buffer_t *msg);
+int buffer_get_ret(buffer_t *msg, void *buf, size_t len);
+int buffer_get_int_ret(unsigned int *ret, buffer_t *msg);
+unsigned int buffer_get_int(buffer_t *msg);
 int buffer_get_char_ret(char *ret, buffer_t *msg);
 int buffer_get_char(buffer_t *msg);
 void buffer_rewind(buffer_t *buf);
